+++ src/contexts/GlobalContext.js
@@ -1,0 +1,3 @@
+import {createContext} from 'react'
+
+export const GlobalContext = createContext()
